fixed logging. elevator module.
