+++ bellboy/utils/messages.py
@@ -1,6 +1,10 @@
 from enum import Enum
 
 """ Global messages, non-actor specifc """
+# init
+class Init:
+    pass
+
 
 # general requests
 class Request(Enum):
@@ -16,8 +20,7 @@ class Response(Enum):
 """ Sensor related messages. """
 # sensor requests
 class SensorReq(Enum):
-    SETUP, CLOSE, POLL = range(3)
-
+    SETUP, CLOSE, POLL, STOP = range(4)
 
 # sensor responses
 class SensorResp(Enum):
@@ -43,6 +46,10 @@ class SensorReqMsg:
         self.pulseWidth_us = pulseWidth_us
         self.sensorEventFunc = triggerFunc
         self.pollPeriod_ms = pollPeriod_ms
+    
+    def __str__(self):
+        return self.type.name
+
 
 
 # sensor events
@@ -55,3 +62,6 @@ class SensorEventMsg:
     def __init__(self, eventType, eventData):
         self.eventType = eventType
         self.eventData = eventData
+
+    def __str__(self):
+        return self.eventType.name
